+++ CHANGES.md
@@ -4,6 +4,8 @@ unreleased
 - Fix a bug in `Ast_builder.Default.pexp_apply` where it would duplicate
   attributes and mess up the locations. (#643. @Skepfyr)
 
+- Add support for OCaml 5.6 primitive aliases (#642, @NathanReb)
+
 0.38.0
 ------